WIP: Support registering values
Values are a tree of objects/arrays with leaf nodes of strings/numbers
based off of what JSON is comprised of. You can register a value at any
path which is pulled apart to form the set of keys that must be
traversed to move across the tree to get to the leaf.

As it stands, there is no real implementation yet. Just a bunch of code
that is starting to solidify my thoughts about how this should be
implemented. Along with a bunch of comments that outline my general
thinking about how things should work.